The damn thing is functional!
It isn't polished and the code tends to be on the level of Spit-and-Bailingwire. But its working and I can use to to set up and water plants!

+++ flora_II/handheld.cpp
@@ -144,9 +144,9 @@ void handheld::handleSetPump(byte* comPtr) {
 
 void handheld::handleSetDryLimit(byte* comPtr) {
 
-  int newVal;
+  byte newVal;
   
-  newVal = *((int*)&(comPtr[1])); // Grab the value from the buffer.
+  newVal = *((byte*)&(comPtr[1])); // Grab the value from the buffer.
   ourParamObj.setDryLimit(newVal);
   comPtr[0] = noErr;
   replyComBuff(1);
